Change how G factor is calculated
There is a race condition where LSTM prediction is done before the H
factors for the previous outputs, leading to differing deque lengths for
beta_N and H89.

This commit changes how G factor is calculated, so that the length of G
is the length of the shortest operand. This is technically incorrect,
but the correctness of G is not critical since it is not put into any of
the NN models.